backend/requirements.txt
# Django and REST Framework
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-extensions>=3.2.3
django-jazzmin>=3.0.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Authentication
PyJWT>=2.8.0
cryptography>=41.0.0

# Production
whitenoise>=6.6.0
gunicorn>=21.2.0

# Monitoring & Error Tracking
sentry-sdk[django]>=1.40.0

# Caching (Production)
django-redis>=5.4.0

# Development
python-dotenv>=1.0.0

# Invoice & PDF Generation
xhtml2pdf>=0.2.11

# Async Task Queue
celery>=5.3.0
redis>=5.0.0

# Supabase Storage Client
supabase>=2.0.0
